Move data types into a matching input module
Now there's matching `inputs` and `outputs` modules.
